requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
PyYAML==6.0.1

# Database
mysql-connector-python==8.2.0
redis==5.0.1

# Telegram
telethon==1.34.0
python-telegram-bot==13.15
PySocks==1.7.1

# HTTP Client
requests==2.31.0
httpx==0.25.2

# Async support
# asyncio is built-in Python 3.4+

# Development (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.12.1
# flake8==6.1.0
